pyproject.toml
[project]
name = "project-guide"
version = "2.0.0"
description = "Stay organized and in control with adaptive LLM workflow prompts."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Pointmatic", email = "pointmatic@example.com" }
]
keywords = [
    "llm",
    "coding",
    "human-in-the-loop",
    "hitloop",
    "documentation",
    "workflow",
    "guide",
    "templates",
    "python",
    "cli",
    "version-management",
    "synchronization",
    "project-management",
    "development-tools",
    "best-practices",
    "override-support",
    "yaml-config",
    "cross-platform"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "packaging>=24.0",
]

[project.urls]
Homepage = "https://github.com/pointmatic/project-guide"
Documentation = "https://pointmatic.github.io/project-guide/"
Repository = "https://github.com/pointmatic/project-guide"
Issues = "https://github.com/pointmatic/project-guide/issues"
Changelog = "https://github.com/pointmatic/project-guide/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]
docs = [
    "mkdocs-material>=9.0",
    "mkdocs-git-revision-date-localized-plugin>=1.0",
]

[project.scripts]
project-guide = "project_guide.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["project_guide"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B904",  # raise from - too strict for our error handling
    "SIM102", # nested if - sometimes clearer as is
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
addopts = "--cov-fail-under=85"
